+++ resources/bluima_chunker_genia/src/main/java/ch/epfl/bbp/nlp/res/chunk/GeniaResource.java
@@ -1,0 +1,13 @@
+package ch.epfl.bbp.nlp.res.chunk;
+
+import ch.epfl.bbp.nlp.ModelResource;
+
+/**
+ * Genia Resource
+ */
+public class GeniaResource implements ModelResource {
+    public String getResourcePath() {
+        return "/Chunker_Genia.bin.gz";
+    }
+}
+
