--- resources/bluima_chunker_genia/src/main/java/ch/epfl/bpp/nlp/res/chunk/GeniaResource.java
@@ -1,13 +1,0 @@
-package ch.epfl.bpp.nlp.res.chunk;
-
-import ch.epfl.bbp.nlp.ModelResource;
-
-/**
- * Genia Resource
- */
-public class GeniaResource implements ModelResource {
-    public String getResourcePath() {
-        return "/Chunker_Genia.bin.gz";
-    }
-}
-
