+++ resources/bluima_postagger_genia/src/main/java/ch/epfl/bbp/nlp/res/tag/GeniaResource.java
@@ -1,0 +1,13 @@
+package ch.epfl.bbp.nlp.res.tag;
+
+import ch.epfl.bbp.nlp.ModelResource;
+
+/**
+ * Genia Resource
+ */
+public class GeniaResource implements ModelResource {
+    public String getResourcePath() {
+        return "/Tagger_Genia.bin.gz";
+    }
+}
+
