+++ resources/bluima_tagdict_genia/src/main/java/ch/epfl/bbp/nlp/res/tag/dict/GeniaResource.java
@@ -1,0 +1,13 @@
+package ch.epfl.bbp.nlp.res.tag.dict;
+
+import ch.epfl.bbp.nlp.ModelResource;
+
+/**
+ * Genia Resource
+ */
+public class GeniaResource implements ModelResource {
+    public String getResourcePath() {
+        return "/tagdict-genia";
+    }
+}
+
