--- resources/bluima_tagdict_genia/src/main/java/ch/epfl/bpp/nlp/res/tag/dict/GeniaResource.java
@@ -1,13 +1,0 @@
-package ch.epfl.bpp.nlp.res.tag.dict;
-
-import ch.epfl.bbp.nlp.ModelResource;
-
-/**
- * Genia Resource
- */
-public class GeniaResource implements ModelResource {
-    public String getResourcePath() {
-        return "/tagdict-genia";
-    }
-}
-
