+++ resources/bluima_tokenizer_genia/src/main/java/ch/epfl/bbp/nlp/res/token/GeniaResource.java
@@ -1,0 +1,13 @@
+package ch.epfl.bbp.nlp.res.token;
+
+import ch.epfl.bbp.nlp.ModelResource;
+
+/**
+ * Genia Resource
+ */
+public class GeniaResource implements ModelResource {
+    public String getResourcePath() {
+        return "/TokenizerGenia.bin.gz";
+    }
+}
+
