--- resources/bluima_tokenizer_genia/src/main/java/ch/epfl/bpp/nlp/res/token/GeniaResource.java
@@ -1,13 +1,0 @@
-package ch.epfl.bpp.nlp.res.token;
-
-import ch.epfl.bbp.nlp.ModelResource;
-
-/**
- * Genia Resource
- */
-public class GeniaResource implements ModelResource {
-    public String getResourcePath() {
-        return "/TokenizerGenia.bin.gz";
-    }
-}
-
